refs: #27 最新情報セクションのコンポーネント化
